requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.3.0
Authlib==1.5.1
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bibtexparser==1.4.3
black==24.10.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
dparse==0.6.4
ecdsa==0.19.1
email-validator==2.1.0.post1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fake-useragent==2.1.0
fastapi==0.109.2
filelock==3.16.1
flake8==3.9.2
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.16.4
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
lxml==5.3.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.6.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.72.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
psutil==6.1.1
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pyflakes==2.3.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.23.8
pytest-cov==2.12.1
pytest-mock==3.14.0
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
rbo==0.1.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.3.7
safetensors==0.5.3
safety==3.3.1
safety-schemas==0.0.11
scholarly==1.7.11
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools-scm==8.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.27
starlette==0.36.3
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.30.2
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.27.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.19.0
